WIP before refactoring plugins
