+++ .config/nvim/lua/config/matchup.lua
@@ -1,0 +1,6 @@
+vim.g.matchup_matchparen_deferred = 1
+vim.g.matchup_matchparen_deferred_show_delay = 100
+vim.g.matchup_matchparen_hi_surround_always = 1
+vim.g.matchup_override_vimtex = 1
+vim.g.matchup_delim_start_plaintext = 0
+vim.g.matchup_transmute_enabled = 0
