+++ .config/nvim/lua/config/signature.lua
@@ -1,0 +1,23 @@
+vim.g.SignatureMap = {
+  -- Leader = 'm',
+  -- PlaceNextMark = '',
+  -- ToggleMarkAtLine = 'm.',
+  -- PurgeMarksAtLine = 'm-',
+  -- DeleteMark = 'dm',
+  -- PurgeMarks = 'm<Space>',
+  -- PurgeMarkers = 'm<BS>',
+  -- GotoNextLineAlpha = '',
+  -- GotoPrevLineAlpha = '',
+  -- GotoNextSpotAlpha = '',
+  -- GotoPrevSpotAlpha = '',
+  -- GotoNextLineByPos = '',
+  -- GotoPrevLineByPos = '',
+  -- GotoNextSpotByPos = '',
+  -- GotoPrevSpotByPos = '',
+  -- GotoNextMarker = '',
+  -- GotoPrevMarker = '',
+  -- GotoNextMarkerAny = '',
+  -- GotoPrevMarkerAny = '',
+  -- ListBufferMarks = 'm/',
+  -- ListBufferMarkers = 'm?'
+}
