+++ .config/nvim/vim/config.vim
