+++ bin/update_all
@@ -5,5 +5,5 @@ cd ~/.oh-my-zsh/custom/plugins/
 find . -type d -depth 1 -exec git --git-dir={}/.git --work-tree=$PWD/{} pull origin master \;
 cd ~/.oh-my-zsh/custom/themes/
 find . -type d -depth 1 -exec git --git-dir={}/.git --work-tree=$PWD/{} pull origin master \;
-nvim +PlugUpdate +qall
+nvim --headless -c 'autocmd User PackerComplete quitall' -c 'PackerSync'
 brew update
